+++ internal/infrastructure/botclient/client_test.go
@@ -19,7 +19,7 @@ var linkUpdate = &dto.LinkUpdate{
 	ID:          1,
 	URL:         "stackoverflow.com",
 	Description: "Тема с вопросом обновилась",
-	TgChatIDs:   []int{2, 8, 12, 15},
+	TgChatIDs:   []int64{2, 8, 12, 15},
 }
 
 func TestBotClient_SendLinkUpdates(t *testing.T) {
